requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.1
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.42.0
Pillow==10.1.0
python-dotenv==1.0.0
rapidfuzz==3.5.2
telethon==1.33.0
apscheduler==3.10.4